requirements.txt
python-dotenv==0.20.0
aiohttp==3.9.5
pyTelegramBotAPI==4.14.1
SQLAlchemy==2.0.31
